refactor: Int 로 받아서 Double 로 계산하도록 수정

+++ src/main/kotlin/lottery/domain/Profit.kt
@@ -1,8 +1,8 @@
 package lottery.domain
 
 object Profit {
-    fun calculate(inputMoney: Double, jackpots: Double): String {
-        return (jackpots / inputMoney).format(2)
+    fun calculate(inputMoney: Int, jackpots: Int): String {
+        return (jackpots.toDouble() / inputMoney.toDouble()).format(2)
     }
 
     private fun Double.format(digits: Int) = "%.${digits}f".format(this)
